Refactor: Moved queries out of page templates and into own file
This lets us abstract the query to keep our staticProps functions tidy
and also allows us to reuse the queries elsewhere.
