+++ src/lib/queryBooks.js
@@ -1,0 +1,177 @@
+import getSanityContent from './sanity';
+// TODO: Write the tests
+
+const querySingleBook = async (slug) => {
+  try {
+    const BookBySlug = await getSanityContent({
+      query: `
+        query BookBySlug($slug: String!) {
+          allBook(where: { slug: { current: { eq: $slug } } }) {
+            title
+            author {
+              name
+            }
+            series {
+              name
+            }
+            blurbRaw
+            cover {
+              asset {
+                url
+                altText
+                metadata {
+                  lqip
+                }
+              }
+            }
+            readStatus
+          }
+        }
+      `,
+      variables: {
+        slug,
+      },
+    });
+
+    const [book] = BookBySlug.allBook;
+
+    return book;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+/**
+ * Query all of the books from the CMS
+ *
+ * @returns {Promise} A Promise containing an object of our query result
+ */
+const queryAllBooks = async () => {
+  try {
+    const AllBooks = await getSanityContent({
+      query: `
+        query AllBooks {
+          allBook {
+            _id
+            title
+            author {
+              name
+            }
+            slug {
+              current
+            }
+            series {
+              name
+            }
+            bookNumber
+            readStatus
+            cover {
+              asset {
+                url
+                altText
+                metadata {
+                  lqip
+                }
+              }
+            }
+          }
+        }
+      `,
+    });
+
+    return AllBooks;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+/**
+ * Query all of our books by the series
+ *
+ * @param {String} variable Title of the series we want to filter our query by
+ * @returns {Promise} A Promise containing an object of our query result
+ */
+const queryBooksBySeries = async (series) => {
+  try {
+    const BooksBySeries = await getSanityContent({
+      query: `
+          query BooksBySeries($series: String!) {
+            allBook(
+              where: { series: { name: { eq: $series } } },
+              sort: { bookNumber: ASC }
+            ) {
+              _id
+              title
+              slug {
+                current
+              }
+              readStatus
+              cover {
+                asset {
+                  url
+                  altText
+                  metadata {
+                    lqip
+                  }
+                }
+              }
+            }
+          }
+        `,
+      variables: {
+        series,
+      },
+    });
+
+    return BooksBySeries;
+  } catch (error) {
+    console.error(error);
+  }
+};
+
+/**
+ *
+ * @param {Array} allBooks
+ */
+const buildBooksArray = async (allBooks) => {
+  const books = await Promise.all(
+    allBooks
+      .filter((book) => book.bookNumber === 1 || book.bookNumber === null)
+      .map(async (book) => {
+        const cover = book.cover ? book.cover : null;
+        const series = book.series ? book.series.name : null;
+
+        // Get books by series and pass them into each response
+        const querySeries = await queryBooksBySeries(series);
+        const relatedBooks = querySeries ? querySeries.allBook : null;
+
+        return {
+          _id: book._id,
+          title: book.title,
+          author: book.author.name,
+          slug: book.slug,
+          series,
+          bookNumber: book.bookNumber,
+          readStatus: book.readStatus,
+          cover,
+          relatedBooks,
+        };
+      }),
+  );
+
+  books.sort((a, b) => {
+    // Sort author alphabetically
+    // If the first item comes first in the alphabet, move it up
+    // Otherwise move it down
+    if (a.author > b.author) return 1;
+    if (a.author < b.author) return -1;
+
+    // Then sort by title
+    if (a.title > b.title) return 1;
+    if (a.title < b.title) return -1;
+  });
+
+  return books;
+};
+
+export { querySingleBook, queryAllBooks, queryBooksBySeries, buildBooksArray };
